Adding <sparkle:tags> appcast item element, <sparkle:criticalUpdate /> tag
You would use this like:
<item>
    <title>My great update</title>
    ...
    <sparkle:tags>
        <sparkle:criticalUpdate />
    </sparkle:tags>
</item>

The behavior of this tag presently is: rather than waiting for the user to quit the app (or for some large amount of time to pass), the user will be prompted to install the update (with a stronger message) immediately upon its receipt. This behavior could be improved for the manual update case as well.

We can also use this new mechanism to create special behavior for e.g. paid updates.